Revision requests took into account.
