+++ Security/TwoFactor/Provider/Google/GoogleAuthenticator.php
@@ -24,6 +24,9 @@ class GoogleAuthenticator implements GoogleAuthenticatorInterface
         $this->issuer = $issuer;
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function checkCode(TwoFactorInterface $user, string $code): bool
     {
         $totp = $this->createTotp($user);
@@ -31,6 +34,9 @@ class GoogleAuthenticator implements GoogleAuthenticatorInterface
         return $totp->verify($code);
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getUrl(TwoFactorInterface $user): string
     {
         $totp = $this->createTotp($user);
@@ -38,6 +44,9 @@ class GoogleAuthenticator implements GoogleAuthenticatorInterface
         return $totp->getQrCodeUri();
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getQRContent(TwoFactorInterface $user): string
     {
         $totp = $this->createTotp($user);
@@ -45,11 +54,19 @@ class GoogleAuthenticator implements GoogleAuthenticatorInterface
         return $totp->getProvisioningUri();
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function generateSecret(): string
     {
         return trim(Base32::encodeUpper(random_bytes(32)), '=');
     }
 
+    /**
+     * @param TwoFactorInterface $user
+     *
+     * @return TOTP
+     */
     private function createTotp(TwoFactorInterface $user): TOTP
     {
         $totp = TOTP::create($user->getGoogleAuthenticatorSecret());
